Refactor API Departments

+++ src/controllers/Department.controller.ts
@@ -2,7 +2,9 @@ import { Request, Response } from 'express';
 import { DepartmentService } from '../services';
 import { formatResponse } from '../utils/response.util';
 import Department from '../models/Department.model';
-import { DepartmentMessages } from '../constants/messages';
+import { AuthMessages, DepartmentMessages } from '../constants/messages';
+import { IDepartments } from '../interfaces/Departments.interface';
+import { Role } from '../constants/enum';
 
 const departmentService = new DepartmentService();
 
@@ -16,18 +18,20 @@ const departmentService = new DepartmentService();
 export const getAllDepartments = async (req: Request, res: Response) => {
   try {
     const {
+      q,
       page = 1,
       limit = 10,
       sortBy = 'createdAt',
       sortOrder = 'asc'
     } = req.query as {
+      q: string
       page?: string;
       limit?: string;
       sortBy?: 'createdAt' | 'name' | 'email';
       sortOrder?: 'asc' | 'desc';
     };
 
-    const result = await departmentService.getDepartments({
+    const result = await departmentService.getDepartments(q, {
       page: parseInt(page as string),
       limit: parseInt(limit as string),
       sortBy,
@@ -61,37 +65,37 @@ export const getAllDepartments = async (req: Request, res: Response) => {
  * @param query 
  * @returns departments
  */
-export const search = async (req: Request, res: Response) => {
-  try {
-    const { q, page = 1, limit = 10 } = req.query as {
-      q: string;
-      page?: string;
-      limit?: string;
-    };
+// export const search = async (req: Request, res: Response) => {
+//   try {
+//     const { q, page = 1, limit = 10 } = req.query as {
+//       q: string;
+//       page?: string;
+//       limit?: string;
+//     };
 
-    const result = await departmentService.search(q, {
-      page: parseInt(page as string),
-      limit: parseInt(limit as string)
-    });
+//     const result = await departmentService.search(q, {
+//       page: parseInt(page as string),
+//       limit: parseInt(limit as string)
+//     });
 
-    res.status(200).json(
-      formatResponse(
-        'success',
-        DepartmentMessages.SEARCH_DEPARTMENT,
-        result.departments,
-        undefined,
-        result.pagination
-      )
-    ); 
-  } catch (error) {
-    res.status(500).json(
-      formatResponse(
-        'error',
-        `Lỗi hệ thống: ${error}`
-      )
-    )
-  }
-}
+//     res.status(200).json(
+//       formatResponse(
+//         'success',
+//         DepartmentMessages.SEARCH_DEPARTMENT,
+//         result.departments,
+//         undefined,
+//         result.pagination
+//       )
+//     ); 
+//   } catch (error) {
+//     res.status(500).json(
+//       formatResponse(
+//         'error',
+//         `Lỗi hệ thống: ${error}`
+//       )
+//     )
+//   }
+// }
 
 /**
  * Function get detail department
@@ -134,7 +138,7 @@ export const createDepartment = async(req: Request, res: Response) => {
     });
     
     if(departmentExit) {
-      res.status(500).json(
+      return res.status(400).json(
         formatResponse(
           'error',
           DepartmentMessages.DEPARTMENT_EXITS
@@ -142,9 +146,13 @@ export const createDepartment = async(req: Request, res: Response) => {
       )
     }
 
-    const department = await departmentService.create(req.body);
+    const dataCreate = {
+      ...req.body,
+      createdBy: req.user?._id
+    } as IDepartments;
+    const department = await departmentService.create(dataCreate);
     
-    res.status(201).json(
+    return res.status(201).json(
       formatResponse(
         'success',
         DepartmentMessages.CREATE_SUCCESSFULLY,
@@ -152,7 +160,7 @@ export const createDepartment = async(req: Request, res: Response) => {
       )
     )
   } catch (error) {
-    res.status(500).json(
+    return res.status(500).json(
       formatResponse(
         'error',
         `Lỗi hệ thống: ${error}`
@@ -170,9 +178,10 @@ export const createDepartment = async(req: Request, res: Response) => {
  */
 export const updateDepartment = async(req: Request, res: Response) => {
   try {
-    const departmentExit = await Department.findById(req.params.id.toString());
+    const currentUser = req.user;
+    const departmentExit = await departmentService.getDepartmentById(req.params.id.toString());
     if(!departmentExit) {
-      res.status(500).json(
+      return res.status(400).json(
         formatResponse(
           'error',
           DepartmentMessages.NOT_FOUND
@@ -180,12 +189,25 @@ export const updateDepartment = async(req: Request, res: Response) => {
       );
     };
 
+    if (currentUser?.role !== Role.ADMIN && departmentExit.createdBy.toString() !== currentUser?._id.toString()) {
+      return res.status(403).json(
+        formatResponse(
+          'error',
+          AuthMessages.FORBIDDEN
+        )
+      );
+    }
+
+    const dataUpdate = {
+      ...req.body,
+      updatedBy: req.user?._id
+    } as IDepartments;
     const department = await departmentService.update(
       req.params.id.toString(),
-      req.body
+      dataUpdate
     );
 
-    res.status(200).json(
+    return res.status(200).json(
       formatResponse(
         'success',
         DepartmentMessages.UPDATE_SUCCESSFULLY,
@@ -193,7 +215,7 @@ export const updateDepartment = async(req: Request, res: Response) => {
       )
     );
   } catch (error) {
-    res.status(500).json(
+    return res.status(500).json(
       formatResponse(
         'error',
         `Lỗi hệ thống: ${error}`
@@ -210,9 +232,10 @@ export const updateDepartment = async(req: Request, res: Response) => {
  */
 export const deleteDepartment = async(req: Request, res: Response) => {
   try {
-    const departmentExit = await Department.findById(req.params.id.toString());
-    if(!departmentExit) {
-      res.status(500).json(
+    const currentUser = req.user;
+    const department = await departmentService.getDepartmentById(req.params.id.toString());
+    if(!department) {
+      return res.status(400).json(
         formatResponse(
           'error',
           DepartmentMessages.NOT_FOUND
@@ -220,15 +243,24 @@ export const deleteDepartment = async(req: Request, res: Response) => {
       );
     };
 
-    const department = await departmentService.delete(req.params.id.toString());
-    res.status(200).json(
+    if (currentUser?.role !== Role.ADMIN && department.createdBy.toString() !== currentUser?._id.toString()) {
+      return res.status(403).json(
+        formatResponse(
+          'error',
+          AuthMessages.FORBIDDEN
+        )
+      );
+    }
+
+    const result = await departmentService.delete(req.params.id.toString());
+    return res.status(200).json(
       formatResponse(
         'success',
         DepartmentMessages.DELETE_SUCCESSFULLY
       )
     );
   } catch (error) {
-    res.status(500).json(
+    return res.status(500).json(
       formatResponse(
         'error',
         `Lỗi hệ thống: ${error}`
